% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAI.R
\name{BAI.P}
\alias{BAI.P}
\title{Basal area increment model Prognaus}
\usage{
BAI.P(
  BART = 1,
  BHD_cm,
  CR_Ant,
  BAL,
  CCF,
  Seehoehe_m = 600,
  Hangneigung_Proz = 50,
  Exposition_Grad = 180,
  HF_cm = 3,
  HH_cm = 2,
  Bodentiefe_cm = 50,
  Bodenfeuchte = 3,
  Relief = 2,
  Bodentyp = 9,
  Vegetationstyp = 1,
  Wuchsbezirk = 13,
  MSE_ja = 0,
  PL = 5
)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{BHD_cm}{dbh in cm}

\item{CR_Ant}{Crown ratio as ratio}

\item{BAL}{Basal area of larger trees (Wykoff, 1990)}

\item{CCF}{Crown competition factor (Krajicek, 1961)}

\item{Seehoehe_m}{Elevation in m (Range: 200-2200 m)}

\item{Hangneigung_Proz}{Slope in percent}

\item{Exposition_Grad}{Aspect in degree}

\item{HF_cm}{Thickness of F-humus layer in cm}

\item{HH_cm}{Thickness of H-humus layer in cm}

\item{Bodentiefe_cm}{Soil depth in cm, model uses dummy variable which is 1 for soildepth <=30cm}

\item{Bodenfeuchte}{Soil moisture,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=3 is "moist"}

\item{Relief}{Slope position,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=2 is "middle slope"}

\item{Bodentyp}{Soil type,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=9 is
"Heavy-textured Cambisols and Luvisols derived from moraine material,
 non-calcareous loess, or mudstone"}

\item{Vegetationstyp}{Vegetation type,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=1 is Shade-tolerant herbs types}

\item{Wuchsbezirk}{Growth district,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=13 is Central Alps}

\item{MSE_ja}{Dummy variable, default=0, if 1 half of the model MSE is
added to correct for logarithmic transformation bias}

\item{PL}{Period length in years, default=5 years, 5 years is maximum
recommended time step for the update of BAL and CCF}
}
\description{
\code{BAI.P} calculates the basal area increment in cm²/period
}
\details{
Calculates the basal area increment of Prognaus. The model is based
on Monserud and Sterba (1996); Coefficients are taken from Hasenauer (2000).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of tree species, soil moisture, relief, soil type, vegetation type
and growth district uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
BAI.P(BART=1,BHD_cm=30,CR_Ant = 0.5,BAL=0,CCF=200)
}
