% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\name{BAL.P}
\alias{BAL.P}
\title{Basal area of larger trees}
\usage{
BAL.P(ID_Punkt = 1, ID_Baum, BHD_cm, G_m2ha)
}
\arguments{
\item{ID_Punkt}{Unique ID for each sample plot}

\item{ID_Baum}{Unique ID for each tree within a plot}

\item{BHD_cm}{dbh in cm}

\item{G_m2ha}{represented basal area in m²/ha}
}
\description{
\code{BAL.P} calculates the basal area of larger trees in m²/ha
}
\details{
Calculates the basal area of larger trees according to Wykoff 1990.
}
