% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR.R
\name{CR.P}
\alias{CR.P}
\title{Crown ratio model Prognaus}
\usage{
CR.P(
  BART = 1,
  BHD_cm,
  H_m,
  BAL,
  CCF,
  Seehoehe_m = 600,
  Hangneigung_Proz = 50,
  Exposition_Grad = 180
)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{BHD_cm}{Dbh in cm}

\item{H_m}{Tree height in m}

\item{BAL}{Basal area of larger trees (Wykoff, 1990)}

\item{CCF}{Crown competition factor (Krajicek, 1961)}

\item{Seehoehe_m}{Elevation in m (Range: 200-2200 m)}

\item{Hangneigung_Proz}{Slope in percent}

\item{Exposition_Grad}{Aspect in degree}
}
\description{
\code{CR.P} calculates the crown ratio
}
\details{
CR.P is a static crown ratio model, i.e. it estimates the crown ratio
from tree and site variables. Dynamic crown ratio models predict the
change in crown ratio. The model was developed by
Hasenauer and Monserud (1996). Coefficients are taken from Hasenauer (2000).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of tree species, uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
CR.P(BART=1,BHD_cm=30,H_m=30,BAL=0,CCF=200)
}
