% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Einwuchs.R
\name{Einwuchs.P}
\alias{Einwuchs.P}
\title{Ingrowth model of Prognaus}
\usage{
Einwuchs.P(
  dg_cm,
  Sum_G,
  sum_G10,
  CCF,
  Seehoehe_m = 600,
  Hangneigung_Proz = 50,
  Exposition_Grad = 180,
  J1 = 0,
  J2 = 0,
  Betriebsart = 1,
  Wuchsbezirk = 13,
  Relief = 2,
  Bodentyp = 9,
  Bodenfeuchte = 3,
  Schlagvegetation = 0,
  PL = 5,
  KGW = 10,
  Stoch = 1,
  spec.vec = c(1, 10),
  spec.EW.vec = c(1, 10)
)
}
\arguments{
\item{dg_cm}{Quadratic mean diameter in cm}

\item{sum_G10}{Basal area m²/ha for trees with a dbh>10cm}

\item{CCF}{Crown competition factor (Krajicek, 1961)}

\item{Seehoehe_m}{Elevation in m (Range: 200-2200 m)}

\item{Hangneigung_Proz}{Slope in percent}

\item{Exposition_Grad}{Aspect in degree}

\item{J1}{Presence(1) / absence(0) of regeneration I (height: 0m-1.3m)}

\item{J2}{Presence(1) / absence(0) of regeneration II (height:1.3m-dbh=10.4cm)}

\item{Betriebsart}{operation type
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=1 is "timber forest, high forest"}

\item{Wuchsbezirk}{Growth district,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=13 is "Central Alps"}

\item{Relief}{Slope position,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=2 is "middle slope"
'@param Bodentyp Soil type,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=9 is
"Heavy-textured Cambisols and Luvisols derived from moraine material,
non-calcareous loess, or mudstone"}

\item{Bodenfeuchte}{Soil moisture,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=3 is "moist"}

\item{Schlagvegetation}{Presence(1) / absence(0) of clear cut vegetation}

\item{PL}{Period length in years, default=5 years, 5 years is maximum
recommended time step for the update of BAL and CCF}

\item{KGW}{Percentage attributed to forest in ANFI, default=10, "Plot entirely located in forest"}

\item{Stoch}{Determines if the model is applied stochastically (1) or deterministically (0).
default=stochastic. The option stochastic/deterministic applies to the model of ingrowth potential
and the Jugend II model. Tree species is alwasy determined stochastically.}

\item{spec.vec}{Vector of tree species that are present on a plot. Tree species,
coding is according to the Austrian National Forest inventory.
\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
By default Norway spruce and European beech are assumed to be present on the plot.}

\item{spec.EW.vec}{Vector of tree species that are allowed to grow into the plot. Tree species,
coding is according to the Austrian National Forest inventory.
\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
By default Norway spruce and European beech are assumed to grow in. Note that 13 tree species models are
available from Ledermann (2002), whereas almost 30 tree species are available in the Austrian National
Forest Inventory. Thus models for some species might not be available and
some models refer to tree species groups.}

\item{sum_G}{Basal area m²/ha}
}
\description{
\code{Einwuchs.P} calculates the ingrowing trees for a plot; If no ingrowth
is predicted the function returns "0", otherwise a matrix of ingrowing trees.
Needs to be applied for each plot
}
\details{
Calculates the ingrowing trees of a plot. The routine calls the following sub-models:
\code{\link{JugendII.P}}: a model to calculate the presence/absence of regeneration II
\code{\link{EWP.P}}: a model to calculate the presence/absence of ingrowth
\code{\link{EW.Anzahl.P}}: a model to calculate the number of ingrowing trees, if ingrowth is present
\code{\link{EinwuchsBA.P}}: a model to determine the tree species of ingrowing trees
\code{\link{EW.BHD.P}}: a model to determine the diameter of ingrowing trees
\code{\link{EWH.P}}: a model to determine the height of ingrowing trees
All models were developed by Ledermann (2002).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of operation type, growth district and soil type uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
Einwuchs.P(dg_cm=25, Sum_G=40, sum_G10=38, CCF=200)


}
