% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Einwuchs.R
\name{EinwuchsBA.P}
\alias{EinwuchsBA.P}
\title{Probability of tree species}
\usage{
EinwuchsBA.P(
  BART = 1,
  Sum_G,
  CCF,
  Seehoehe_m = 600,
  Hangneigung_Proz = 50,
  Exposition_Grad = 180,
  Bodenfeuchte = 3,
  Wuchsbezirk = 13,
  Bodentyp = 9,
  Betriebsart = 1,
  spec.vec = c(1, 10)
)
}
\arguments{
\item{BART}{Tree species for which the probability of ingrowth is predicted
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce#'}

\item{CCF}{Crown competition factor (Krajicek, 1961)}

\item{Seehoehe_m}{Elevation in m (Range: 200-2200 m)}

\item{Hangneigung_Proz}{Slope in percent}

\item{Exposition_Grad}{Aspect in degree}

\item{Bodenfeuchte}{Soil moisture,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=3 is "moist"}

\item{Wuchsbezirk}{Growth district,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=13 is "Central Alps"}

\item{Bodentyp}{Soil type,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=9 is
"Heavy-textured Cambisols and Luvisols derived from moraine material,
non-calcareous loess, or mudstone"}

\item{Betriebsart}{operation type
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=1 is "timber forest, high forest"}

\item{sum_G}{Basal area m²/ha}

\item{species.vec}{Vector with species present on plot}
}
\description{
\code{EinwuchsBA} calculates the probability of presence/absence for a certain tree species
}
\details{
Calculates the probability of presence/absence for a certain tree species.
During the simulation the caluculation needs to be done for each of the 13 tree species or tree species groups
and for the number of ingrowing trees predicted by \code{\link{EW.Anzahl.P}}. The choice of tree species is
determined by the call in the ingrowth routine.
The model is part ot the modelling system
for calculating ingrowth on plots developed by Ledermann (2002).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of tree species, soil moisture, growth district, soil type and operation type uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
EinwuchsBA.P(Sum_G=40,  CCF=200)
}
