% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Einwuchs.R
\name{EW.Anzahl.P}
\alias{EW.Anzahl.P}
\title{Number of ingrowing trees on a plot}
\usage{
EW.Anzahl.P(
  dg_cm,
  Sum_G,
  CCF,
  Seehoehe_m = 600,
  Hangneigung_Proz = 50,
  Exposition_Grad = 180,
  J1 = 0,
  J2 = 0,
  Betriebsart = 1,
  Wuchsbezirk = 13,
  PL = 5,
  KGW = 10
)
}
\arguments{
\item{dg_cm}{Quadratic mean diameter in cm}

\item{CCF}{Crown competition factor (Krajicek, 1961)}

\item{Seehoehe_m}{Elevation in m (Range: 200-2200 m)}

\item{Hangneigung_Proz}{Slope in percent}

\item{Exposition_Grad}{Aspect in degree}

\item{J1}{Presence(1) / absence(0) of regeneration I (height: 0m-1.3m)}

\item{J2}{Presence(1) / absence(0) of regeneration II (height:1.3m-dbh=10.4cm)}

\item{Betriebsart}{operation type
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=1 is "timber forest, high forest"}

\item{Wuchsbezirk}{Growth district,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=13 is "Central Alps"}

\item{PL}{Period length in years, default=5 years, 5 years is maximum
recommended time step for the update of BAL and CCF}

\item{KGW}{Percentage attributed to forest in ANFI, default=10, "Plot entirely located in forest"}

\item{sum_G}{Basal area m²/ha}
}
\description{
\code{EW.Anzahl.P} calculates the number of ingrowing trees on a plot
in a 5-year period
}
\details{
Calculates the number of ingrowing trees on a plot, if there is presence of ingrowth predicted by
\code{\link{EWP.P}}. The model is part ot the modelling system
for calculating ingrowth on plots developed by Ledermann (2002).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of operation type and growth district uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
EW.Anzahl.P(dg_cm=30,  Sum_G=40,  CCF=200)
}
