% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Einwuchs.R
\name{EWH.P}
\alias{EWH.P}
\title{Height of ingrowing trees for Prognaus}
\usage{
EWH.P(
  BART_EW = 1,
  BHD_EW_cm,
  Seehoehe_m = 600,
  dg_cm,
  Sum_G,
  sum_G10,
  CCF,
  Relief = 2
)
}
\arguments{
\item{BART_EW}{Tree species of ingrowing tree,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{BHD_EW_cm}{Diameter of ingrowing tree in cm}

\item{Seehoehe_m}{Elevation in m (Range: 200-2200 m)}

\item{dg_cm}{Quadratic mean diameter in cm}

\item{Sum_G}{Basal area m²/ha}

\item{CCF}{Crown competition factor (Krajicek, 1961)}

\item{Relief}{Slope position,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=2 is "middle slope"}

\item{Sum_G10}{Basal area od trees with a diamter > 10 cm, in m²/ha}
}
\description{
\code{EWH.P} calculates the height for ingrowing trees
}
\details{
Calculates the height of ingrowing trees in m. The model is part ot the modelling system
for calculating ingrowth on plots developed by Ledermann (2002).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of tree species of ingrowing trees and relief uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
EWH.P(BHD_EW_cm=6,  dg_cm=10,
Sum_G=40,  sum_G10=38,  CCF=200,  Relief=2)
}
