% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\name{EZ}
\alias{EZ}
\title{Open-grown tree crown area}
\usage{
EZ(SD_m, nrep = 1)
}
\arguments{
\item{SD_m}{Diameter of an open-grown tree in meter}

\item{nrep}{represented stem number, default=1, data is not from
a sample, i.e. each tree represents 1 tree per hectare}
}
\description{
\code{EZ} calculates the represented open-grown tree crown area
in m²
}
\details{
Calculates the represented open-grown crown area from an open-grown tree
diameter \code{\link{SD.P}} assuming circular crown areas and multiplies
them by the represented stem number (e.g. in fixed area plots or
angle count samples)
}
