% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mort.R
\name{GMax}
\alias{GMax}
\title{Maximum basal area functions}
\usage{
GMax(BART = 1, dg_cm)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{dg_cm}{Quadratic mean diameter in cm}
}
\description{
\code{GMax} calculates the maximum basal area according to the Reineke-rule
}
\details{
GMax calculates the maximum basal area with 0.95 quantile coefficients according to Vospernik (2015)
#'For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
}
\examples{
GMax(BART=1,dg_cm=30)
}
