% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HI.R
\name{ih.P}
\alias{ih.P}
\title{Height increment model Prognaus}
\usage{
ih.P(
  BART = 1,
  H_m,
  Ho_m,
  BAL,
  CCF,
  Seehoehe_m = 600,
  Hangneigung_Proz = 50,
  Exposition_Grad = 180,
  Bodentiefe_cm = 50,
  Bodenfeuchte = 3,
  Relief = 2,
  Bodentyp = 9,
  Wuchsbezirk = 13,
  PL = 5
)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{H_m}{Tree height in m}

\item{Ho_m}{Dominant height in m}

\item{BAL}{Basal area of larger trees (Wykoff, 1990)}

\item{CCF}{Crown competition factor (Krajicek, 1961)}

\item{Seehoehe_m}{Elevation in m (Range: 200-2200 m)}

\item{Hangneigung_Proz}{Slope in percent}

\item{Exposition_Grad}{Aspect in degree}

\item{Bodentiefe_cm}{Soil depth in cm, model uses dummy variable which is 1 for soildepth <=30cm}

\item{Bodenfeuchte}{Soil moisture,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=3 is "moist"}

\item{Relief}{Slope position,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=2 is "middle slope"}

\item{Bodentyp}{Soil type,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=9 is
"Heavy-textured Cambisols and Luvisols derived from moraine material,
non-calcareous loess, or mudstone"}

\item{Wuchsbezirk}{Growth district,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default}

\item{PL}{Period length in years, default=5 years, 5 years is maximum
recommended time step for the update of BAL and CCF}
}
\description{
\code{ih.P} calculates the height increment in m/period
}
\details{
Calculates the height increment of Prognaus. The model is an Evolon-model
(Mende and Albrecht, 2001). The model is published by Nachtmann (2005).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of tree species, soil moisture, relief, soil type, vegetation type
and growth district uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
ih.P(BART=1,H_m=30,Ho_m=30,BAL=0,CCF=200)

height2012<-30
height2015<-height2012+ih.P(BART=1,H_m=30,Ho_m=30,BAL=0,CCF=200,PL=3)
}
