% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Einwuchs.R
\name{JugendII.P}
\alias{JugendII.P}
\title{Regeneration II model Prognaus}
\usage{
JugendII.P(
  dg_cm,
  Sum_G,
  CCF,
  Seehoehe_m = 600,
  Hangneigung_Proz = 50,
  J1 = 0,
  J2 = 0,
  Betriebsart = 1,
  Wuchsbezirk = 13,
  Bodentyp = 9,
  Schlagvegetation = 0,
  PL = 5,
  KGW = 10,
  Stoch = 1
)
}
\arguments{
\item{dg_cm}{Quadratic mean diameter in cm}

\item{CCF}{Crown competition factor (Krajicek, 1961)}

\item{Seehoehe_m}{Elevation in m (Range: 200-2200 m)}

\item{Hangneigung_Proz}{Slope in percent}

\item{J1}{Presence(1) / absence(0) of regeneration I (height: 0m-1.3m)}

\item{J2}{Presence(1) / absence(0) of regeneration II (height:1.3m-dbh=10.4cm)}

\item{Betriebsart}{operation type
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=1 is "timber forest, high forest"}

\item{Wuchsbezirk}{Growth district,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=13 is Central Alps
'@param Bodentyp Soil type,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}, default=9 is
"Heavy-textured Cambisols and Luvisols derived from moraine material,
 non-calcareous loess, or mudstone"}

\item{Schlagvegetation}{Presence(1) / absence(0) of clear cut vegetation}

\item{PL}{Period length in years, default=5 years, 5 years is maximum
recommended time step for the update of BAL and CCF}

\item{KGW}{Percentage attributed to forest in ANFI, default=10, "Plot entirely located in forest"}

\item{sum_G}{Basal area m²/ha}
}
\description{
\code{JugendII.P} calculates the occurrence of regeneration II on a plot (TRUE/FALSE)
}
\details{
Calculates if regeneration II occurs on the plot (presence/absence). The model is part ot the modelling system
for calculating ingrowth on plots developed by Ledermann (2002).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of operation type, growth district and soil type uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
JugendII.P(dg_cm=30,  Sum_G=40,  CCF=200)
}
