% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mort.R
\name{Mort.P}
\alias{Mort.P}
\title{Mortality Model Prognaus}
\usage{
Mort.P(BART = 1, BHD_cm, BAL = 0, CR_Ant = 0.5, PL = 5)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{BHD_cm}{Dbh in cm#}

\item{BAL}{Basal area of larger trees (Wykoff, 1990)}

\item{CR}{Crown ratio as ratio}
}
\description{
\code{Mort.P} calculates mortality rates
}
\details{
Mort.P calculates the Mortality rates of a logistic model.
The model was developed by Monserud and Sterba (1999).
Coefficients are taken from Hasenauer (2000).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of tree species, uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
Mort.P(BART=1,BHD_cm=30,BAL=0,CR_Ant=0.5)

#Deterministic
nrep2012<-500
nrep2015<-nrep2012*(1-Mort.P(BART=1,BHD_cm=30,BAL=0,CR_Ant=0.5,PL=3))

#Stochastic
ZZ<-runif(n=1,min=0,max=1)
ifelse(Mort.P(BART=1,BHD_cm=30,BAL=0,CR_Ant=0.5)>ZZ,1,0)
}
