% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OH.R
\name{OH.Ass}
\alias{OH.Ass}
\title{Dominant height according to Assmann}
\usage{
OH.Ass(ID_Punkt = 1, BART = 1, BHD_cm, Hoehe_m, nrep)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{BHD_cm}{dbh in cm}

\item{nrep}{represented stem number, default=1, data is not from
a sample, i.e. each tree represents 1 tree per hectare}

\item{ID}{Unique ID for each sample plot}
}
\description{
\code{OH.Ass} calculates the dominant height
of 100 largest trees per hectare in m
}
\details{
Calculates the dominant height according to Assmann as Lorey's mean
height, dominant height is calculated independent of the tree species.
Missing dominant heights are imputed using the quotients between tree
species from Vospernik (2000).
Function not suitable for full cruise!!!!
}
