% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp.R
\name{SD.P}
\alias{SD.P}
\title{Open grown tree crown diameter}
\usage{
SD.P(BART = 1, BHD_cm, MSE_ja = 0)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{BHD_cm}{dbh in cm}

\item{MSE_ja}{Dummy variable, default=0, if 1 half of the model MSE is
added to correct for logarithmic transformation bias}
}
\description{
\code{SD.P} calculates the crown diameter of open-grown trees in m
}
\details{
The model is based on Hasenauer (1997).
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}.
Coding of tree species uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
