% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomasse.R
\name{Spez.Gewicht}
\alias{Spez.Gewicht}
\title{Specific weight}
\usage{
Spez.Gewicht(BART = 1)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}
}
\description{
\code{Spez.Gewicht} calculates the specific weight in kg/m³
}
\details{
Calculates the specific weight according to usage of trait in Austria ÖHU(2006)
}
\examples{
Spez.Gewicht(BART=1)

}
