% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPoll.R
\name{vPoll}
\alias{vPoll}
\title{Individual tree volume equation}
\usage{
vPoll(BART, BHD_cm, H_m)
}
\arguments{
\item{BART}{Tree species,
coding=\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI},
default=1 Norway spruce}

\item{BHD_cm}{Dbh in cm}

\item{H_m}{Tree height in m}
}
\description{
\code{vPoll} calculates the individual tree volume in m³
}
\details{
The caclulation is based on form factor equation according
to Pollanschütz (1974) \code{\link{fPoll}}.
For full references see
\url{https://homepage.boku.ac.at/sonja/Lit_Prog.pdf}
Coding of tree species uses the coding of the
Austrian National Forest Inventory
(\href{https://homepage.boku.ac.at/sonja/ANFI_E.pdf}{ANFI}).
}
\examples{
vPoll(BART=1,BHD_cm=30,H_m=30)
}
